SOL-151648: Split loader tests into focused files
